+++ lib/src/anim.dart
@@ -2,17 +2,25 @@ import 'package:boomsheets/src/keyframe.dart';
 import 'package:boomsheets/src/frametime.dart';
 import 'package:yes_parser/yes_parser.dart';
 
+/// [Anim] represents a single animation state with a [name].
+///
+/// [totalDuration] calculates the total number of frames as [Frametime] units.
+
 class Anim {
   Frametime? _totalDuration;
   List<Attribute> attrs;
   List<Keyframe> keyframes;
   final String name;
 
-  /// Creates state [Anim] called [name] with [keyframes] and [attributes].
-  Anim(this.name, {List<Keyframe>? keyframes, List<Attribute>? attributes})
-      : keyframes = keyframes ?? [],
+  /// Creates state [Anim] called [name] with provided data.
+  Anim(
+    this.name, {
+    List<Keyframe>? keyframes,
+    List<Attribute>? attributes,
+  })  : keyframes = keyframes ?? [],
         attrs = attributes ?? [];
 
+  // Calculate and caches the total duration of this anim state.
   Frametime get totalDuration {
     if (_totalDuration != null) return _totalDuration!;