+++ lib/src/labeled_point.dart
@@ -1,7 +1,11 @@
 import 'dart:math';
 
+/// Point<int> [pos] with a String [label].
 class LabeledPoint {
   String label;
   Point<int> pos;
   LabeledPoint({required this.label, required this.pos});
+
+  @override
+  String toString() => "point={$label, x=${pos.x}, y=${pos.y}}";
 }